+++ inst/paramtest/test_paramtest_classif.glmnet.R
@@ -15,8 +15,7 @@ test_that("classif.glmnet", {
     "family", # handled by mlr3
     "itrace", # supported via param trace.it
     "factory", # only used in scripts, no effect within mlr3
-    "offset", # handled by mlr3
-    "sigma2.threshold" # added by mlr3
+    "offset" # handled by mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "train")
@@ -36,8 +35,7 @@ test_that("predict classif.glmnet", {
     "newx", # handled via mlr3
     "type", # handled via mlr3
     "newoffset", # handled via mlr3
-    "use_pred_offset", # handled via mlr3
-    "sigma2.threshold" # added by mlr3
+    "use_pred_offset" # handled via mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "predict")